+++ plugin-animation/src/main/java/com/mapbox/maps/plugin/animation/CameraAnimationsPluginImpl.kt
@@ -11,6 +11,7 @@ import androidx.annotation.VisibleForTesting.PRIVATE
 import com.mapbox.geojson.Point
 import com.mapbox.maps.*
 import com.mapbox.maps.plugin.animation.animator.*
+import com.mapbox.maps.plugin.animation.threads.MapboxAnimatorThread
 import com.mapbox.maps.plugin.delegates.*
 import com.mapbox.maps.util.MathUtils
 import java.util.concurrent.CopyOnWriteArraySet