+++ plugin-locationcomponent/src/test/java/com/mapbox/maps/plugin/locationcomponent/animators/PuckAnimatorManagerTest.kt
@@ -4,7 +4,7 @@ import android.animation.ValueAnimator
 import android.graphics.Color
 import android.os.Looper
 import com.mapbox.geojson.Point
-import com.mapbox.maps.plugin.animation.MapboxAnimatorThread
+import com.mapbox.maps.plugin.animation.threads.MapboxAnimatorThread
 import com.mapbox.maps.plugin.locationcomponent.LocationLayerRenderer
 import com.mapbox.maps.plugin.locationcomponent.generated.LocationComponentSettings
 import com.mapbox.maps.plugin.locationcomponent.generated.LocationComponentSettings2